Added reports Added authorization in ApiController Added RedisRepositories Added some javadocs Other minor changes

+++ src/main/java/com/glackfag/shorty/repositories/jpa/ReportRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ReportRepository extends JpaRepository<Report, Integer> {
-    boolean existsByAliasAndReporterId(String alias, int reporterId);
+    boolean existsByAliasAndReporterId(String alias, long reporterId);
 }
